Feat: Adapter, Alien-base using new Masterchef
