+++ src/adapters/magpie/common/penpie.ts
@@ -1,7 +1,7 @@
 import { getPendlePools } from '@adapters/pendle/common/pool'
 import type { BaseContext, Contract } from '@lib/adapter'
 import { mapSuccessFilter, rangeBI } from '@lib/array'
-import { getMasterChefPoolsContracts, type GetPoolsInfosParams } from '@lib/masterchef/newMasterchef'
+import { getMasterChefPoolsContracts, type GetPoolsInfosParams } from '@lib/masterchef/masterChefContract'
 import { multicall } from '@lib/multicall'
 
 const abi = {
