+++ src/adapters/mdex/bsc/index.ts
@@ -1,6 +1,7 @@
 import type { BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
 import { resolveBalances } from '@lib/balance'
-import { getMasterChefPoolsBalances, getMasterChefPoolsContracts } from '@lib/masterchef/newMasterchef'
+import { getMasterChefPoolsBalances } from '@lib/masterchef/masterChefBalance'
+import { getMasterChefPoolsContracts } from '@lib/masterchef/masterChefContract'
 import type { Token } from '@lib/token'
 import { getPairsContracts } from '@lib/uniswap/v2/factory'
 import { getPairsBalances } from '@lib/uniswap/v2/pair'
